test(modal): split card tests for paralellization
